bigger bullets

+++ ArcadeGame/src/core/Bullet.java
@@ -16,7 +16,7 @@ public class Bullet extends Collider
 		this.owner = owner;
 		this.setMobile(true);
 	}
-	public Bullet(double x, double y, String owner, String image,int width,int height,int damage) {
+	public Bullet(double x, double y, String owner, String image,double width,double height,int damage) {
 		super(new String[] {image}, 1, x, y, width, height, 0, 0);
 		this.owner = owner;
 		this.damage = damage;
